feat(dx): rename __main__.py to cli.py and adapt imports
This change was needed to make `uv run verbatim` work
