+++ run.py
@@ -1,7 +1,20 @@
 import sys
-from verbatim.__main__ import main
+from verbatim.cli import main
 
-sys.argv = ['run.py', 'samples/voices.wav', '-w', 'out', '--language', 'en', 'fr', '--txt', '--md', '--json', '--docx', '--ass']
+sys.argv = [
+    "run.py",
+    "samples/voices.wav",
+    "-w",
+    "out",
+    "--language",
+    "en",
+    "fr",
+    "--txt",
+    "--md",
+    "--json",
+    "--docx",
+    "--ass",
+]
 
 # Call the main function
 main()
