+++ verbatim/cli.py
